+++ FluentEmail/IMustAddBody.cs
@@ -1,3 +1,4 @@
+using System.Net.Mime;
 using System.Text;
 
 namespace FluentEmail
@@ -6,5 +7,7 @@ namespace FluentEmail
     {
         ICanAddAttachmentOrBuild Body(string body);
         ICanAddAttachmentOrBuild Body(string body, Encoding encodingType);
+        ICanAddAttachmentOrBuild Body(string body, TransferEncoding transferEncoding);
+        ICanAddAttachmentOrBuild Body(string body, Encoding encodingType, TransferEncoding transferEncoding);
     }
 }